Split support functions to multiple files.
